separate map and bar chart code to prep for automating 4 markets
