+++ gulp/config.js
@@ -64,8 +64,8 @@ module.exports = {
   scripts: {
     src: [
       loc.src + '/static/js/metadata.js',
-      loc.src + '/static/js/auto-loan-charts.js',
-      loc.src + '/static/js/auto-loan-charts.js',
+      loc.src + '/static/js/bar.js',
+      loc.src + '/static/js/tile-map.js',
       loc.src + '/static/js/formatDates.js',
       loc.src + '/static/js/templates/charts.js',
       loc.src + '/static/js/utils/getFilePath.js',
