+++ gulp/tasks/watch.js
@@ -10,7 +10,7 @@ var connect = require( 'gulp-connect' );
 var config = require( '../config' );
 
 gulp.task( 'watch', [ 'connect', 'browserSync' ], function() {
-  gulp.watch( config.scripts.src, [ 'scripts', 'handlebars' ] );
+  gulp.watch( config.scripts.src, [ 'handlebars' ] );
   gulp.watch( config.handlebarsTemplates.dom, [ 'release' ] )
   gulp.watch( config.handlebarsTemplates.charts, [ 'handlebars' ] )
   gulp.watch( config.handlebarsTemplates.src, [ 'handlebars' ] )
