+++ src/static/css/charts.less
@@ -26,6 +26,12 @@
 
     }
 
+    &[data-chart-type=map] {
+        h3 {
+            display: none;
+        }
+    } 
+
     .line {
       fill: none;
       stroke: steelblue;
